server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Uncomment for MongoDB support (see server/app/db.py)
# motor>=3.3.0
